+++ packages/components/src/components/common/input-base/input-base-styles.ts
@@ -23,6 +23,7 @@ import type { FormState } from '../../../utils/form/form-state';
 import { getFunctionalComponentLabelAfterStyles, getFunctionalComponentLabelStyles } from '../label/label-styles';
 import { getFunctionalComponentLoadingMessageStyles } from '../loading-message/loading-message-styles';
 import { getFunctionalComponentStateMessageStyles } from '../state-message/state-message-styles';
+import { mergeDeep } from '../../../utils';
 
 /**
  * @css-variable {"name": "--ref-p-input-slotted-padding", "description": "When slotting a `p-button-pure` or `p-link-pure` this variable needs to be set as `padding` in oder to adjust the alignment correctly."}
@@ -62,14 +63,15 @@ export const getFunctionalComponentInputBaseStyles = (
           [`${cssVarButtonPurePadding}`]: buttonPadding,
           [`${cssVarButtonPureMargin}`]: buttonMargin,
           ...hostHiddenStyles,
-          ...(isDisabled && getDisabledBaseStyles()),
         }),
       },
-      ...getFunctionalComponentLabelAfterStyles(isDisabled),
+      ...getFunctionalComponentLabelAfterStyles(),
       ...preventFoucOfNestedElementsStyles,
       input: {
         all: 'unset',
+        display: 'flex',
         flex: 1,
+        alignItems: 'center',
         width: 'max(100%, 2ch)', // show at least 2 characters in very narrow containers
         height: '100%',
         font: textSmallStyle.font.replace('ex', 'ex + 6px'), // a minimum line-height is needed for input, otherwise value is scrollable in Chrome, +6px is aligned with how Safari visualize date/time input highlighting
@@ -106,10 +108,13 @@ export const getFunctionalComponentInputBaseStyles = (
           outlineOffset: '2px',
         }),
       },
-      ...(isDisabled &&
-        forcedColorsMediaQuery({
-          borderColor: 'GrayText',
-        })),
+      ...(isDisabled && {
+        ...mergeDeep({ ...getDisabledBaseStyles() }, {
+          ...forcedColorsMediaQuery({
+            borderColor: 'GrayText',
+          }),
+        }),
+      }),
       ...(!isDisabled &&
         !readOnly &&
         !isLoading &&
@@ -127,7 +132,7 @@ export const getFunctionalComponentInputBaseStyles = (
       },
     }),
     // .label / .required
-    ...getFunctionalComponentLabelStyles(isDisabled, hideLabel),
+    ...getFunctionalComponentLabelStyles(isDisabled, isLoading, hideLabel),
     // .message
     ...getFunctionalComponentStateMessageStyles(state),
     // .loading