+++ packages/components/src/components/radio-group/radio-group-option/radio-group-option-styles.ts
@@ -52,10 +52,9 @@ export const getComponentCss = (disabled: boolean, loading: boolean, state: Radi
         display: 'block',
         ...addImportantToEachRule({
           ...hostHiddenStyles,
-          ...(disabledOrLoading && getDisabledBaseStyles()),
         }),
       },
-      ...getFunctionalComponentLabelAfterStyles(disabledOrLoading),
+      ...getFunctionalComponentLabelAfterStyles(),
       ...preventFoucOfNestedElementsStyles,
       input: {
         all: 'unset',
@@ -114,6 +113,7 @@ export const getComponentCss = (disabled: boolean, loading: boolean, state: Radi
       alignSelf: 'flex-start',
       minHeight: fontLineHeight, // necessary for compact mode
       cursor: disabledOrLoading ? 'not-allowed' : 'pointer',
+      ...(disabled && getDisabledBaseStyles()),
     },
     ...(loading && {
       spinner: {
@@ -127,7 +127,7 @@ export const getComponentCss = (disabled: boolean, loading: boolean, state: Radi
       },
     }),
     // .label / .required
-    ...getFunctionalComponentLabelStyles(disabled || loading, false, null, {
+    ...getFunctionalComponentLabelStyles(disabled, loading, false, null, {
       paddingTop: labelPaddingTop,
       paddingInlineStart: labelPaddingInlineStart,
     }),