+++ packages/storefront/src/app/components/input-email/input-email.stories.ts
@@ -79,7 +79,7 @@ export const inputEmailSlotStories: SlotStories<'p-input-email'> = {
       generator: () => [
         {
           tag: 'p-popover',
-          properties: { slot: 'label-after', className: 'ms-static-xs' },
+          properties: { slot: 'label-after' },
           children: ['Some Popover Content.'],
         },
       ],
@@ -125,7 +125,7 @@ export const inputEmailStorySlots: Story<'p-input-email'> = {
         },
         {
           tag: 'p-popover',
-          properties: { slot: 'label-after', className: 'ms-static-xs' },
+          properties: { slot: 'label-after' },
           children: [
             'Some Popover content with a ',
             {